+++ public/js/post/delete.js
@@ -1,9 +1,13 @@
+const id = document.currentScript.getAttribute('post-id');
+
 document.getElementById('delete').onclick = e => {
-    fetch('/admin/posts/<%= post.id %>', {
-        method: 'DELETE',
-    }).then(response => {
-        if (response.status === 200) {
-            window.location.href = '/';
-        }
-    });
+    if (confirm("Are you sure you want to delete this post?")) {
+        fetch(`/admin/posts/${id}`, {
+            method: 'DELETE',
+        }).then(response => {
+            if (response.status === 200) {
+                window.location.href = '/';
+            }
+        });
+    }
 };
